iteration 4 and 5 ready, except normalization

+++ db/migrate/20170320222114_create_trips.rb
@@ -1,16 +1,16 @@
 class CreateTrips < ActiveRecord::Migration[5.0]
   def change
     create_table :trips do |t|
-      t.integer :duration
+      t.bigint :duration
       t.date :start_date
       t.text :start_station_name
-      t.integer :start_station_id
+      t.bigint :start_station_id
       t.date :end_date
       t.text :end_station_name
-      t.integer :end_station_id
+      t.bigint :end_station_id
       t.integer :bike_id
       t.text :subscription_type
-      t.integer :zip_code
+      t.bigint :zip_code
       t.timestamps null: false
     end
   end
